Created Context

+++ src/server.rs
@@ -1,10 +1,8 @@
-use crate::MqttServer;
-use crate::WebServer;
-use daemonize::Daemonize;
+use crate::context::Context;
+use crate::{log, MqttServer, WebServer};
 use nix::sys::signal::{kill, Signal};
 use nix::unistd::Pid;
 use std::fs;
-use std::io::Write;
 use std::process::exit;
 use tokio::signal::unix::{signal, SignalKind};
 use tracing::{error, info};
@@ -39,10 +37,16 @@ impl Server {
                 }
             }
 
-            let daemon =
-                Daemonize::new().pid_file(PID_FILE).chown_pid_file(true).working_directory("./");
+            let stdout = fs::File::create("/tmp/daemon.out").unwrap();
+            let stderr = fs::File::create("/tmp/daemon.err").unwrap();
+            let daemon = daemonize::Daemonize::new()
+                .pid_file(PID_FILE)
+                .chown_pid_file(true)
+                .working_directory("./")
+                .stdout(stdout)
+                .stderr(stderr);
             if let Err(e) = daemon.start() {
-                error!("Daemonize: {}", e);
+                eprintln!("Daemonize failed: {}", e);
                 exit(1);
             }
         }
@@ -54,26 +58,29 @@ impl Server {
 
     // stop server
     pub fn stop() {
-        match fs::read_to_string(PID_FILE) {
-            Ok(pid) => {
-                if let Ok(pid) = pid.trim().parse::<i32>() {
-                    if let Err(e) = kill(Pid::from_raw(pid), Signal::SIGTERM) {
-                        println!("{} [PID: {}]", e, pid);
-                        let _ = fs::remove_file(PID_FILE);
-                    }
+        if let Ok(pid) = fs::read_to_string(PID_FILE) {
+            if let Ok(pid) = pid.trim().parse::<i32>() {
+                if let Err(e) = kill(Pid::from_raw(pid), Signal::SIGTERM) {
+                    println!("{} [PID: {}]", e, pid);
+                    let _ = fs::remove_file(PID_FILE);
                 }
-            }
-            Err(_) => {
-                println!("{} server is not running", env!("CARGO_PKG_NAME"));
             }
         }
     }
 
+    fn is_stop() -> bool {
+        !std::path::Path::new(PID_FILE).exists()
+    }
+
     // restart server
     pub fn restart() {
-        info!("{} server restarting...", env!("CARGO_PKG_NAME"));
         Self::stop();
-        std::thread::sleep(std::time::Duration::from_secs(1));
+        for _ in 0..30 {
+            if Self::is_stop() {
+                break;
+            }
+            std::thread::sleep(std::time::Duration::from_millis(100));
+        }
         Self::start();
     }
 
@@ -86,21 +93,22 @@ impl Server {
 
 // run server
 async fn run() {
-    let (shutdown_tx, _) = tokio::sync::broadcast::channel::<()>(1);
+    log::init();
+    let ctx = Context::new().build();
 
     // Web Server
-    let rx = shutdown_tx.subscribe();
+    let web_ctx = ctx.clone();
     let web_task = tokio::spawn(async move {
-        if let Err(e) = WebServer::start(rx).await {
+        if let Err(e) = WebServer::start(web_ctx).await {
             error!("WebServer: {}", e);
             exit(1);
         }
     });
 
     // Mqtt Server
-    let rx = shutdown_tx.subscribe();
+    let mqtt_ctx = ctx.clone();
     let mqtt_task = tokio::spawn(async move {
-        if let Err(e) = MqttServer::start(rx).await {
+        if let Err(e) = MqttServer::start(mqtt_ctx).await {
             error!("MqttServer: {}", e);
             exit(1);
         }
@@ -113,23 +121,22 @@ async fn run() {
     loop {
         tokio::select! {
             _ = sigint.recv() => {
-                info!("Server received SIGINT signal");
-                let  _ = shutdown_tx.send(());
+                info!("server received SIGINT signal");
+                let  _ = ctx.shutdown();
                 break;
             }
             _ = sigterm.recv() => {
-                info!("Server received SIGTERM signal");
-                let  _ = shutdown_tx.send(());
+                info!("server received SIGTERM signal");
+                let  _ = ctx.shutdown();
                 break;
             }
             _ = sighup.recv() => {
-                info!("Server received SIGHUP signal");
+                info!("server received SIGHUP signal");
                 Server::reload();
             }
         }
     }
 
     let _ = tokio::join!(web_task, mqtt_task);
-
     let _ = fs::remove_file(PID_FILE);
 }
